3_wf_3_function-speechsense-upload/requirements.txt
grpcio>=1.64.1
grpcio-tools>=1.64.1
numpy>=2.0.0
yandexcloud>=0.295.0
boto3>=1.34.138
psycopg2>=2.9.9
protobuf>=5.27.2
jsonschema>=4.23.0